chatbot-educatif-backend/requirements.txt
Django==5.1.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.0
python-decouple==3.8
PyMySQL==1.1.2
PyJWT==2.11.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0